+++ src/commands/use.js
@@ -8,6 +8,7 @@ const { extractSection } = require('../core/section');
 const { PROMPT_MAX_CHARS } = require('../utils/constants');
 const { initPlugins } = require('../plugins/loader');
 const { addTask } = require('../utils/task-history');
+const { evaluatePromptBudget } = require('../utils/token-estimator');
 
 const COMPACT_THRESHOLD = PROMPT_MAX_CHARS;
 const LOW_CONFIDENCE_THRESHOLD = 50;
@@ -224,7 +225,8 @@ async function useCommand(options = {}) {
     compactInfo,
     matchMethod,
     aiReason,
-    loadedDocs
+    loadedDocs,
+    tokenBudget: evaluatePromptBudget(prompt, aiConfig?.maxTokens)
   };
 }